+++ test/upgrader_test.dart
@@ -131,12 +131,14 @@ void main() {
     expect(upgrader.messages!.buttonTitleIgnore, 'IGNORE');
     expect(upgrader.messages!.buttonTitleLater, 'LATER');
     expect(upgrader.messages!.buttonTitleUpdate, 'UPDATE NOW');
+    expect(upgrader.messages!.releaseNotes, 'Release Notes');
 
     upgrader.messages = MyUpgraderMessages();
 
     expect(upgrader.messages!.buttonTitleIgnore, 'aaa');
     expect(upgrader.messages!.buttonTitleLater, 'bbb');
     expect(upgrader.messages!.buttonTitleUpdate, 'ccc');
+    expect(upgrader.messages!.releaseNotes, 'ddd');
 
     await tester.pumpWidget(const _MyWidget());
 
@@ -151,19 +153,21 @@ void main() {
 
     expect(find.text(upgrader.messages!.title), findsOneWidget);
     expect(find.text(upgrader.message()), findsOneWidget);
-    expect(find.text('Release Notes:'), findsOneWidget);
+    expect(find.text(upgrader.messages!.releaseNotes), findsOneWidget);
     expect(find.text(upgrader.releaseNotes!), findsOneWidget);
     expect(find.text(upgrader.messages!.prompt), findsOneWidget);
     expect(find.byType(TextButton), findsNWidgets(3));
     expect(find.text(upgrader.messages!.buttonTitleIgnore), findsOneWidget);
     expect(find.text(upgrader.messages!.buttonTitleLater), findsOneWidget);
     expect(find.text(upgrader.messages!.buttonTitleUpdate), findsOneWidget);
+    expect(find.text(upgrader.messages!.releaseNotes), findsOneWidget);
 
     await tester.tap(find.text(upgrader.messages!.buttonTitleUpdate));
     await tester.pumpAndSettle();
     expect(find.text(upgrader.messages!.buttonTitleIgnore), findsNothing);
     expect(find.text(upgrader.messages!.buttonTitleLater), findsNothing);
     expect(find.text(upgrader.messages!.buttonTitleUpdate), findsNothing);
+    expect(find.text(upgrader.messages!.releaseNotes), findsNothing);
     expect(called, true);
     expect(notCalled, true);
   }, skip: false);
@@ -228,7 +232,7 @@ void main() {
 
     expect(find.text(upgrader.messages!.title), findsOneWidget);
     expect(find.text(upgrader.message()), findsOneWidget);
-    expect(find.text('Release Notes:'), findsOneWidget);
+    expect(find.text(upgrader.messages!.releaseNotes), findsOneWidget);
     expect(find.text(upgrader.releaseNotes!), findsOneWidget);
     expect(find.text(upgrader.messages!.prompt), findsOneWidget);
     expect(find.byType(CupertinoDialogAction), findsNWidgets(3));
@@ -406,7 +410,7 @@ void main() {
     // Pump the UI so the upgrade card is displayed
     await tester.pumpAndSettle();
 
-    expect(find.text('Release Notes:'), findsOneWidget);
+    expect(find.text(upgrader.messages!.releaseNotes), findsOneWidget);
     expect(find.text(upgrader.releaseNotes!), findsOneWidget);
     await tester.tap(find.text(upgrader.messages!.buttonTitleUpdate));
     await tester.pumpAndSettle();
@@ -709,6 +713,24 @@ void main() {
       expect(upgrader.shouldDisplayUpgrade(), true);
       upgrader.debugDisplayAlways = false;
       expect(upgrader.shouldDisplayUpgrade(), false);
+
+      // Test the willDisplayUpgrade callback
+      var notCalled = true;
+      upgrader.willDisplayUpgrade = (bool value) {
+        expect(value, false);
+        notCalled = false;
+      };
+      expect(upgrader.shouldDisplayUpgrade(), false);
+      expect(notCalled, false);
+
+      upgrader.debugDisplayAlways = true;
+      notCalled = true;
+      upgrader.willDisplayUpgrade = (bool value) {
+        expect(value, true);
+        notCalled = false;
+      };
+      expect(upgrader.shouldDisplayUpgrade(), true);
+      expect(notCalled, false);
     }, skip: false);
 
     test('should return true when version is below minAppVersion', () async {
@@ -789,10 +811,12 @@ void main() {
     verifyMessages(UpgraderMessages(code: 'hu'), 'hu');
     verifyMessages(UpgraderMessages(code: 'id'), 'id');
     verifyMessages(UpgraderMessages(code: 'it'), 'it');
+    verifyMessages(UpgraderMessages(code: 'ja'), 'ja');
     verifyMessages(UpgraderMessages(code: 'kk'), 'kk');
     verifyMessages(UpgraderMessages(code: 'km'), 'km');
     verifyMessages(UpgraderMessages(code: 'ko'), 'ko');
     verifyMessages(UpgraderMessages(code: 'lt'), 'lt');
+    verifyMessages(UpgraderMessages(code: 'mn'), 'mn');
     verifyMessages(UpgraderMessages(code: 'nb'), 'nb');
     verifyMessages(UpgraderMessages(code: 'nl'), 'nl');
     verifyMessages(UpgraderMessages(code: 'pt'), 'pt');
@@ -837,6 +861,7 @@ class _MyWidget extends StatelessWidget {
         body: UpgradeAlert(
             debugLogging: true,
             dialogStyle: dialogStyle,
+            willDisplayUpgrade: (bool value) {},
             child: Column(
               children: const <Widget>[Text('Upgrading')],
             )),
@@ -859,7 +884,12 @@ class _MyWidgetCard extends StatelessWidget {
           title: const Text('Upgrader test'),
         ),
         body: Column(
-          children: <Widget>[UpgradeCard(debugLogging: true)],
+          children: <Widget>[
+            UpgradeCard(
+              debugLogging: true,
+              willDisplayUpgrade: (bool value) {},
+            )
+          ],
         ),
       ),
     );
@@ -873,4 +903,6 @@ class MyUpgraderMessages extends UpgraderMessages {
   String get buttonTitleLater => 'bbb';
   @override
   String get buttonTitleUpdate => 'ccc';
+  @override
+  String get releaseNotes => 'ddd';
 }